+++ python/lsst/ts/tcpip/one_client_server.py
@@ -29,7 +29,12 @@ import typing
 
 from . import utils
 from .base_client_or_server import BaseClientOrServer, ConnectCallbackType
-from .constants import DEFAULT_MONITOR_CONNECTION_INTERVAL
+from .constants import (
+    DEFAULT_ENCODING,
+    DEFAULT_LOCALHOST,
+    DEFAULT_MONITOR_CONNECTION_INTERVAL,
+    DEFAULT_TERMINATOR,
+)
 
 
 class OneClientServer(BaseClientOrServer):
@@ -39,18 +44,17 @@ class OneClientServer(BaseClientOrServer):
 
     Parameters
     ----------
-    host : `str` or `None`
-        IP address for this server; typically `LOCALHOST` to get
-        the default version of IP, or `LOCALHOST_IPV4` for IPV4,
-        or `LOCALHOST_IPV6` for IPV6.
-        If `None` then bind to all network interfaces
-        (e.g. listen on an IPv4 socket and an IPv6 socket).
-        Warning: `None` can cause trouble with ``port=0``; see ``port``
-        in the Attributes section for more information.
     port : `int`
         IP port for this server. If 0 then randomly pick an available port
         (or ports, if listening on multiple sockets).
         0 is strongly recommended for unit tests.
+    host : `str` or `None`
+        IP address for this server. The default is `DEFAULT_LOCALHOST`.
+        Specify `LOCALHOST_IPV4` to force IPV4 or `LOCALHOST_IPV6` for IPV6.
+        If `None` then bind to all network interfaces
+        (e.g. listen on an IPv4 socket and an IPv6 socket).
+        Warning: `None` can cause trouble with ``port=0``; see ``port``
+        in the Attributes section for more information.
     log : `logging.Logger`
         Logger.
     connect_callback : callable or `None`, optional
@@ -68,6 +72,12 @@ class OneClientServer(BaseClientOrServer):
         to detect and report hangups).
     name : `str`, optional
         Name used for log messages, e.g. "Commands" or "Telemetry".
+    encoding : `str`
+        The encoding used by `read_str` and `write_str`, `read_json`,
+         and `write_json`.
+    terminator : `bytes`
+        The terminator used by `read_str` and `write_str`, `read_json`,
+         and `write_json`.
     **kwargs : `dict` [`str`, `typing.Any`]
         Additional keyword arguments for `asyncio.start_server`,
         beyond host and port.
@@ -122,12 +132,15 @@ class OneClientServer(BaseClientOrServer):
 
     def __init__(
         self,
-        host: str | None,
+        *,
         port: int | None,
+        host: str | None = DEFAULT_LOCALHOST,
         log: logging.Logger,
         connect_callback: ConnectCallbackType | None = None,
         monitor_connection_interval: float = DEFAULT_MONITOR_CONNECTION_INTERVAL,
         name: str = "",
+        encoding: str = DEFAULT_ENCODING,
+        terminator: bytes = DEFAULT_TERMINATOR,
         **kwargs: typing.Any,
     ) -> None:
         self.host = host
@@ -139,6 +152,8 @@ class OneClientServer(BaseClientOrServer):
             connect_callback=connect_callback,
             monitor_connection_interval=monitor_connection_interval,
             name=name,
+            encoding=encoding,
+            terminator=terminator,
             **kwargs,
         )